pyproject.toml
[tool.poetry]
name = "rp-test"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.2"
colorama = "^0.4.6"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"

[tool.blue]
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
